+++ usermgmt_gui/src/main.rs
@@ -15,7 +15,10 @@ pub mod main_logic;
 fn main() -> Result<(), eframe::Error> {
     // Set up logging and panic messages with link to issue page
     usermgmt_lib::app_panic_hook::set_app_panic_hook();
-    env_logger::init();
+    // Logger handler in variable so background thread for file logging is not stopped until the
+    // end of application.
+    let _keep_logger_handler = usermgmt_lib::logging::set_up_logging(env!("CARGO_PKG_NAME"))
+        .expect("Failed to initilize logger");
 
     // Construct application state before starting the main window for egui frontend.
     // This default impl for app state panics if set up failed due to invalid setting files